requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bson==0.5.8
dataclasses==0.6
Django==2.2.11
django-bootstrap4==1.1.1
django-redis-cache==2.1.0
djangorestframework==3.11.0
djongo==1.3.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
